requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-embed-video==1.1.2
idna==2.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
